Add files via upload
Welcome to the Guessing Number Game! 🎉

In this game, the computer will randomly select a number between a specified range (for example, 1 to 100), and your goal is to guess that number. 

Here's how to play:
1. You will be prompted to enter your guess.
2. After each guess, the game will tell you if your guess is too high, too low, or correct.
3. Keep guessing until you find the correct number!
4. The game will also keep track of the number of attempts you take to guess the correct number.

Are you ready to test your guessing skills? Let’s see how quickly you can find the right number! Good luck! 🍀